requirements.txt
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-http
robotframework >= 7.0